refactor(mantine): reorganize table code
