+++ packages/mantine/src/components/table/Table.tsx
@@ -7,18 +7,18 @@ import {
     flexRender,
     getCoreRowModel,
     Row,
-    TableState,
     useReactTable,
+    TableState,
 } from '@tanstack/react-table';
-import {CoreOptions, InitialTableState, TableOptions} from '@tanstack/table-core';
 import debounce from 'lodash.debounce';
 import defaultsDeep from 'lodash.defaultsdeep';
-import {Children, Fragment, ReactElement, ReactNode, useCallback, useEffect, useState} from 'react';
+import {Children, Fragment, ReactElement, useCallback, useEffect, useState} from 'react';
 
 import useStyles from './Table.styles';
+import {TableFormType, TableProps, TableType} from './Table.types';
 import {TableActions} from './TableActions';
 import {TableAccordionColumn, TableCollapsibleColumn} from './TableCollapsibleColumn';
-import {onTableChangeEvent, TableContext, TableFormType} from './TableContext';
+import {TableContext} from './TableContext';
 import {TableDateRangePicker} from './TableDateRangePicker';
 import {TableFilter} from './TableFilter';
 import {TableFooter} from './TableFooter';
@@ -29,104 +29,6 @@ import {TablePredicate} from './TablePredicate';
 import {TableSelectableColumn} from './TableSelectableColumn';
 import {Th} from './Th';
 import {useRowSelection} from './useRowSelection';
-
-export interface TableProps<T> {
-    /**
-     * Data to display in the table
-     */
-    data: T[];
-    /**
-     * Defines how each row is uniquely identified. It is highly recommended that you specify this prop to an ID that makes sense.
-     */
-    getRowId?: CoreOptions<T>['getRowId'];
-    /**
-     * Columns to display in the table.
-     *
-     * @see https://tanstack.com/table/v8/docs/guide/column-defs
-     */
-    columns: Array<ColumnDef<T>>;
-    /**
-     * Function called when the table mounts
-     *
-     * @param state the state of the table
-     */
-    onMount?: onTableChangeEvent;
-    /**
-     * Function called when the table should update
-     *
-     * @param state the state of the table
-     */
-    onChange?: onTableChangeEvent;
-    /**
-     * Function that generates the expandable content of a row
-     * Return null for rows that don't need to be expandable
-     *
-     * @param datum the row for which the children should be generated.
-     */
-    getExpandChildren?: (datum: T) => ReactNode;
-    /**
-     * React children to show when the table has no rows to show. You can leverage useTable to get the state of the table
-     */
-    noDataChildren?: ReactNode;
-    /**
-     * Whether the table is loading or not
-     *
-     * @default false
-     */
-    loading?: boolean;
-    /**
-     * Childrens to display in the table. They need to be wrap in either `Table.Header` or `Table.Footer`
-     *
-     * @example
-     * <Table ...>
-     *     <Table.Header>
-     *         <div>Hello</div>
-     *     </Table.Header>
-     * </Table>
-     */
-    children?: ReactNode;
-
-    /**
-     * Initial state of the table
-     */
-    initialState?: InitialTableState & Partial<TableFormType>;
-    /**
-     * Action passed when user double clicks on a row
-     */
-    doubleClickAction?: (datum: T) => void;
-    /**
-     * Whether the user can select multiple rows in order to perform actions in bulk
-     *
-     * @default false
-     */
-    multiRowSelectionEnabled?: boolean;
-
-    options?: Omit<
-        Partial<TableOptions<T>>,
-        | 'initialState'
-        | 'data'
-        | 'columns'
-        | 'manualPagination'
-        | 'enableMultiRowSelection'
-        | 'getRowId'
-        | 'getRowCanExpand'
-        | 'enableRowSelection'
-    >;
-}
-
-interface TableType {
-    <T>(props: TableProps<T>): ReactElement;
-    Actions: typeof TableActions;
-    Filter: typeof TableFilter;
-    Footer: typeof TableFooter;
-    Header: typeof TableHeader;
-    Pagination: typeof TablePagination;
-    PerPage: typeof TablePerPage;
-    Predicate: typeof TablePredicate;
-    DateRangePicker: typeof TableDateRangePicker;
-    CollapsibleColumn: typeof TableCollapsibleColumn;
-    AccordionColumn: typeof TableAccordionColumn;
-}
 
 export const Table: TableType = <T,>({
     data,
