--- packages/mantine/src/components/table/useTable.tsx
@@ -1,11 +1,0 @@
-import {useContext} from 'react';
-import {TableContext} from './TableContext';
-
-export const useTable = () => {
-    const ctx = useContext(TableContext);
-    if (ctx === null) {
-        throw new Error('useTable must be used inside of a TableContext.Provider');
-    }
-
-    return ctx;
-};
